Define a QCA vendor command to configure SAR Power limits
There is a regulatory requirement for Specific Absorption Rate (SAR)
whereby the device transmit power is reduced when it is determined that
the device is in close proximity to the body. Implement a vendor command
interface to allow a userspace entity to dynamically control the SAR
power limits.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
Git-commit: c79238b6a460ab6bc6ebc5e2453fd94716393105
Git-repo : git://w1.fi/srv/git/hostap.git
Change-Id: Ieefa5ac2e08df9db95b37c8d250a822313e42f48
CRs-Fixed: 2015906