eap_proxy: Fix eap_proxy_init() prototype to use const eapol_cb
The eapol_cb structure was made const and that change resulted in a
compilation warning/error if CONFIG_EAP_PROXY=<name> is enabled in the
wpa_supplicant build configuration. Fix this by updating the function
prototype to match the change.

Note: This results in a change needed to external eap_proxy_*.c
implementations to match the change.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
CRs-Fixed: 1103504
Change-Id: I022c8415306ec817f80f57c551594480f28d3e09
Git-commit: 79a54ab9f65448a9f3f8f149da5379614ad7bdca
Git-repo : git://w1.fi/srv/git/hostap.git