Use random MAC address for scanning only in non-connected state
cfg80211 rejects the scans issued with random MAC address if the STA is
in connected state. This resulted in failures when using MAC_RAND_SCAN
while connected (CTRL-EVENT-SCAN-FAILED ret=-95). Enable random MAC
address functionality only if the STA is not in connected state to avoid
this. The real MAC address of the STA is already revealed in the
association, so this is an acceptable fallback mechanism for now.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
Git-commit: 346b333d9a57155e5a32a31d951ada38b3c899fc
Git-repo : git://w1.fi/srv/git/hostap.git
CRs-fixed: 1112901

Change-Id: I39bd5608ca0319bac453a889c86c86105e8e0938